more

+++ m2min/otccelf_cc_x86_min.c
@@ -115,8 +115,8 @@ int brk(void *addr)
 	    "int !0x80");
 }
 
-long _malloc_ptr;
-long _brk_ptr;
+int _malloc_ptr;
+int _brk_ptr;
 
 void* malloc(int size)
 {
@@ -132,7 +132,7 @@ void* malloc(int size)
 		if(-1 == _brk_ptr) return 0;
 	}
 
-	long old_malloc = _malloc_ptr;
+	int old_malloc = _malloc_ptr;
 	_malloc_ptr = _malloc_ptr + size;
 	return old_malloc;
 }
